use ‘Maybe’ in the permutation function rather than the magic value ‘ignore’
Fixes AccelerateHS/accelerate#87
