+++ src/Data/Array/Accelerate/Pattern/Maybe.hs
@@ -1,0 +1,26 @@
+{-# LANGUAGE GADTs               #-}
+{-# LANGUAGE PatternSynonyms     #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TemplateHaskell     #-}
+{-# LANGUAGE TypeApplications    #-}
+{-# LANGUAGE ViewPatterns        #-}
+-- |
+-- Module      : Data.Array.Accelerate.Pattern.Maybe
+-- Copyright   : [2018..2019] The Accelerate Team
+-- License     : BSD3
+--
+-- Maintainer  : Trevor L. McDonell <trevor.mcdonell@gmail.com>
+-- Stability   : experimental
+-- Portability : non-portable (GHC extensions)
+--
+
+module Data.Array.Accelerate.Pattern.Maybe (
+
+  Maybe, pattern Nothing_, pattern Just_,
+
+) where
+
+import Data.Array.Accelerate.Pattern.TH
+
+mkPattern ''Maybe
+
